moving navigate tests to beagle as well as newPath variable
